Update
Entire map updated, every object placed perfectly. Only things left to
do are the real estates. Only bug I found which I'm unsure how to fix is
on the right hand side of the map, the hospitals are missing apart from
2 guys and an ATM. All billboards are complete, just need to create a
police image to put in kav police station. Also, there are a few places
that might need a marker or something, like.. when you buy a boat, the
boat appears X (here).