Updated.
Map updated. Kav car marker sorted. Looked into the hospital issue on
the right hand side of the map, still unsure how to solve that, the
hospital buildings are completely missing and they're seemingly
unavailable from the menu, it's a default building.

Also have possible fix's for a few of the bugs listed on the site,
although not 100% sure, so I won't change anything until I verify them
on ts.. just to be safe.

and the cop uniform, theres only 1 so its being added back in, I assume
it needs to be...?